requirements.txt
captum==0.6.0
gdown==4.7.1
GitPython==3.1.30
imbalanced_learn==0.10.1
numpy==1.24.3
pandas==2.0.2
PyDriller==2.4.1
scikit_learn==1.2.2
scipy==1.10.1
tensorflow==2.11.0
tokenizers==0.11.4
torch==1.8.1+cu111
tqdm==4.65.0
transformers==4.24.0
tree_sitter==0.20.1
